# openapi_specs/openapi.yaml
components:
  schemas:
    BoardCreate:
      properties:
        description:
          anyOf:
          - maxLength: 1000
            minLength: 1
            type: string
          - type: 'null'
          title: Description
        memberIds:
          anyOf:
          - items:
              format: uuid
              type: string
            type: array
          - type: 'null'
          default: []
          title: Memberids
        name:
          maxLength: 50
          minLength: 1
          title: Name
          type: string
      required:
      - name
      title: BoardCreate
      type: object
    BoardDetailRead:
      properties:
        createdAt:
          format: date-time
          title: Createdat
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        id:
          format: uuid
          title: Id
          type: string
        memberIds:
          anyOf:
          - items:
              format: uuid
              type: string
            type: array
          - type: 'null'
          default: []
          title: Memberids
        name:
          title: Name
          type: string
        ownerId:
          format: uuid
          title: Ownerid
          type: string
        position:
          title: Position
          type: integer
        tables:
          default: []
          items:
            $ref: '#/components/schemas/TableRead'
          title: Tables
          type: array
        updatedAt:
          format: date-time
          title: Updatedat
          type: string
      required:
      - id
      - name
      - description
      - ownerId
      - position
      - createdAt
      - updatedAt
      title: BoardDetailRead
      type: object
    BoardRead:
      properties:
        createdAt:
          format: date-time
          title: Createdat
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        id:
          format: uuid
          title: Id
          type: string
        memberIds:
          anyOf:
          - items:
              format: uuid
              type: string
            type: array
          - type: 'null'
          default: []
          title: Memberids
        name:
          title: Name
          type: string
        ownerId:
          format: uuid
          title: Ownerid
          type: string
        position:
          title: Position
          type: integer
        updatedAt:
          format: date-time
          title: Updatedat
          type: string
      required:
      - id
      - name
      - description
      - ownerId
      - position
      - createdAt
      - updatedAt
      title: BoardRead
      type: object
    BoardUpdate:
      properties:
        description:
          anyOf:
          - maxLength: 1000
            minLength: 1
            type: string
          - type: 'null'
          title: Description
        name:
          anyOf:
          - maxLength: 50
            minLength: 1
            type: string
          - type: 'null'
          title: Name
        position:
          anyOf:
          - maximum: 100.0
            minimum: 0.0
            type: integer
          - type: 'null'
          title: Position
      title: BoardUpdate
      type: object
    ErrorResponseModel:
      properties:
        error:
          description: Error type
          title: Error
          type: string
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          description: Custom error code or identifier
          title: Error Code
        message:
          description: Human-readable error message
          title: Message
          type: string
      required:
      - error
      - message
      title: ErrorResponseModel
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    PriorityEnum:
      enum:
      - low
      - medium
      - high
      - critical
      title: PriorityEnum
      type: string
...
    RowOwnerRead:
      properties:
        avatarUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatarurl
        email:
          title: Email
          type: string
        firstName:
          title: Firstname
          type: string
        id:
          format: uuid
          title: Id
          type: string
        lastName:
          title: Lastname
          type: string
      required:
      - id
      - firstName
      - lastName
      - email
      title: RowOwnerRead
      type: object
    RowRead:
      properties:
        createdAt:
          format: date-time
          title: Createdat
          type: string
        dueDate:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Duedate
        id:
          format: uuid
          title: Id
          type: string
        name:
          title: Name
          type: string
        owners:
          default: []
          items:
            $ref: '#/components/schemas/RowOwnerRead'
          title: Owners
          type: array
        position:
          title: Position
          type: integer
        priority:
          $ref: '#/components/schemas/PriorityEnum'
        status:
          $ref: '#/components/schemas/StatusEnum'
        tableId:
          format: uuid
          title: Tableid
          type: string
        updatedAt:
          format: date-time
          title: Updatedat
          type: string
      required:
      - id
      - tableId
      - name
      - status
      - priority
      - position
      - createdAt
      - updatedAt
      title: RowRead
      type: object
...
    StatusEnum:
      enum:
      - not_started
      - stuck
      - working_on_it
      - done
      title: StatusEnum
      type: string
...
    TableRead:
      properties:
        boardId:
          description: Foreign key for board
          format: uuid
          title: Boardid
          type: string
        color:
          description: Table color
          title: Color
          type: string
        createdAt:
          description: Table creation date
          format: date-time
          title: Createdat
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          description: Table description
          title: Description
        id:
          description: Primary key for table
          format: uuid
          title: Id
          type: string
        name:
          description: Table name
          title: Name
          type: string
        position:
          description: Table order
          title: Position
          type: integer
        rows:
          description: List of rows in the table
          items:
            $ref: '#/components/schemas/RowRead'
          title: Rows
          type: array
        updatedAt:
          description: Table update date
          format: date-time
          title: Updatedat
          type: string
      required:
      - id
      - boardId
      - name
      - description
      - position
      - color
      - createdAt
      - updatedAt
      - rows
      title: TableRead
      type: object
...
    UserCreate:
      properties:
        email:
          format: email
          maxLength: 255
          minLength: 1
          title: Email
          type: string
        firstName:
          maxLength: 50
          minLength: 1
          title: Firstname
          type: string
        lastName:
          maxLength: 50
          minLength: 1
          title: Lastname
          type: string
        password:
          maxLength: 128
          minLength: 6
          title: Password
          type: string
      required:
      - firstName
      - lastName
      - email
      - password
      title: UserCreate
      type: object
    UserLogin:
      properties:
        email:
          format: email
          title: Email
          type: string
        password:
          title: Password
          type: string
      required:
      - email
      - password
      title: UserLogin
      type: object
    UserRead:
      properties:
        avatarUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatarurl
        createdAt:
          format: date-time
          title: Createdat
          type: string
        email:
          format: email
          title: Email
          type: string
        firstName:
          title: Firstname
          type: string
        id:
          format: uuid
          title: Id
          type: string
        lastName:
          title: Lastname
          type: string
        updatedAt:
          format: date-time
          title: Updatedat
          type: string
      required:
      - id
      - firstName
      - lastName
      - email
      - createdAt
      - updatedAt
      title: UserRead
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: Monday Lite API
  version: 0.1.0
openapi: 3.1.0
paths:
  /api/v1/auth/login:
    post:
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Login
      tags:
      - auth
  /api/v1/auth/logout:
    get:
      operationId: logout
      parameters:
      - in: cookie
        name: refresh_token
        required: false
        schema:
          title: Refresh Token
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Logout
      tags:
      - auth
  /api/v1/auth/refresh-token:
    post:
      operationId: refresh_token
      parameters:
      - in: cookie
        name: refresh_token
        required: false
        schema:
          title: Refresh Token
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Refresh Token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Register
      tags:
      - auth
  /api/v1/boards/:
    get:
      operationId: list_boards
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BoardRead'
                title: Response List Boards Api V1 Boards  Get
                type: array
          description: Successful Response
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: List Boards
      tags:
      - board
    post:
      operationId: create_board
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Create Board
      tags:
      - board
  /api/v1/boards/{board_id}:
    delete:
      operationId: delete_board
      parameters:
      - in: path
        name: board_id
        required: true
        schema:
          format: uuid
          title: Board Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Delete Board
      tags:
      - board
    get:
      operationId: get_board
      parameters:
      - in: path
        name: board_id
        required: true
        schema:
          format: uuid
          title: Board Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardDetailRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Get Board
      tags:
      - board
    patch:
      operationId: update_board
      parameters:
      - in: path
        name: board_id
        required: true
        schema:
          format: uuid
          title: Board Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Update Board
      tags:
      - board
  /api/v1/boards/{board_id}/duplicate:
    post:
      operationId: duplicate_board
      parameters:
      - in: path
        name: board_id
        required: true
        schema:
          format: uuid
          title: Board Id
          type: string
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Duplicate Board
      tags:
      - board
  /api/v1/boards/{board_id}/members/{user_id}:
    delete:
      operationId: remove_member
      parameters:
      - in: path
        name: board_id
        required: true
        schema:
          format: uuid
          title: Board Id
          type: string
      - in: path
        name: user_id
        required: true
        schema:
          format: uuid
          title: User Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Remove Member
      tags:
      - board
    post:
      operationId: add_member
      parameters:
      - in: path
        name: board_id
        required: true
        schema:
          format: uuid
          title: Board Id
          type: string
      - in: path
        name: user_id
        required: true
        schema:
          format: uuid
          title: User Id
          type: string
      responses:
        '201':
          content:
            application/json:
              schema:
                format: uuid
                title: Response Add Member Api V1 Boards  Board Id  Members  User
                  Id  Post
                type: string
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              example:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred. Please try again later.
              schema:
                $ref: '#/components/schemas/ErrorResponseModel'
          description: Internal Server Error
      summary: Add Member
      tags:
      - board
...